Add fighting system between Heroes and Monsters

+++ src/main/kotlin/model/map_character/MapCharacter.kt
@@ -4,7 +4,7 @@ package model.map_character
 Interface with shared behaviour between Monsters and Heroes
  */
 interface MapCharacter {
-    var health: Int
+    var health: Double
 
     //TODO add behaviours shared by all Map Character
     fun attack(opponent: MapCharacter)
